refactor(lib): rename `xframe-reactive` to `xframe`
