+++ Cargo.toml
@@ -1,6 +1,3 @@
-[workspace]
-members = ["packages/reactive"]
-
 [package]
 name = "xframe"
 version = "0.0.0"
@@ -8,4 +5,8 @@ edition = "2021"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
-xframe-reactive = { path = "packages/reactive" }
+ahash = "0.8"
+bumpalo = "3.11"
+indexmap = "1.9"
+smallvec = "1.10"
+slotmap = "1.0"
