--- packages/reactive/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "xframe-reactive"
-version = "0.0.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-ahash = "0.8"
-bumpalo = "3.11"
-indexmap = "1.9"
-smallvec = "1.10"
-slotmap = "1.0"
