+++ src/scope.rs
