+++ src/shared.rs
